Using mmap to reduce memory usage
mmap allows the kernel to swap in and out pages
with file contents on disk. So while the virtual
memory space consumed will still equal the size
of the filesystem, the kernel may use as little
as 1 page of physical memory at a time.

Also added contextlib.ExitStack so that multiple
resources can be released at completion of the
with-statement.