Using memoryview to speed up processing
Previously, the remainder of the filesystem was
copied in memory every time a JFFS2 header was
encountered. For large filesystems, this created a
huge mmap, a userspace copy, and then a munmap.
Most of the newly mmap'd space was not accessed
and the userspace copy greatly increased the
processing time.

Now with the memoryview, the mmap and the
userspace copy are completely avoided with minimal
code changes. The memoryview just points to
offsets within the filesystem, and likewise slices
into the memoryview point to new offsets within
the filesystem memory mapping.