Fix duplicate inodes handling.
When duplicate inodes were found, a new filesystem entry was created in
the fs dict, leading to the creation of hundreds of fs_ directories if
the parsed filesystem had been in used for a while.

To fix this, we simply consider the inode with the highest version value
if a duplicate inode use is detected. This means we keep a single
filesystem output directory, with the most recent version of the DIRENT
there.